backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
requests
httpx
google-generativeai
scikit-learn
pandas
numpy
python-multipart